// apps/api/src/modules/auth/enums/role.enum.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  DIRECTOR = 'DIRECTOR',
  SALES_MANAGER = 'SALES_MANAGER',
  TECHNOLOGIST = 'TECHNOLOGIST',
  FINANCE = 'FINANCE',
  FINANCE_MANAGER = 'FINANCE_MANAGER',
  HR_MANAGER = 'HR_MANAGER',
  PRODUCTION_MANAGER = 'PRODUCTION_MANAGER',
  WAREHOUSE = 'WAREHOUSE',
  MANAGER = 'MANAGER',
}